--- a/scripts/bench/bench-packages.ts
+++ b/scripts/bench/bench-packages.ts
@@ -460,16 +460,6 @@ const run = async () => {
   if (options.baseBranch) {
     const comparisonResults = await compareResults({ results, baseBranch: options.baseBranch });
     const resultsAboveThreshold = filterResultsByThresholds(comparisonResults);
-    await saveLocally({
-      filename: `compare-with-${options.baseBranch}.json`,
-      results: comparisonResults,
-      diff: true,
-    });
-    await saveLocally({
-      filename: `comparisons-above-threshold-with-${options.baseBranch}.json`,
-      results: resultsAboveThreshold,
-      diff: true,
-    });
     if (options.pullRequest) {
       await uploadToGithub({
         results: resultsAboveThreshold,